feat(supply): update supply pickup timing to use seconds and improve message formatting

+++ src/client/java/net/iqaddons/mod/events/impl/skyblock/SupplyPickupEvent.java
@@ -6,6 +6,6 @@ public record SupplyPickupEvent(
         String originalMessage,
         String playerName,
         int currentSupply,
-        long pickupAt
+        double pickupAt
 ) implements Event {
 }
